Add more Ubuntu and Debian awareness or support.

+++ install_clouderadirector.sh
@@ -14,17 +14,63 @@
 #
 # Copyright Clairvoyant 2015
 
-if rpm -q redhat-lsb-core; then
-  OSREL=`lsb_release -rs | awk -F. '{print $1}'`
-else
-  OSREL=`rpm -qf /etc/redhat-release --qf="%{VERSION}\n"`
+# Function to discover basic OS details.
+discover_os () {
+  if command -v lsb_release >/dev/null; then
+    # CentOS, Ubuntu
+    OS=`lsb_release -is`
+    # 7.2.1511, 14.04
+    OSVER=`lsb_release -rs`
+    # 7, 14
+    OSREL=`echo $OSVER | awk -F. '{print $1}'`
+    # trusty, wheezy, Final
+    OSNAME=`lsb_release -cs`
+  else
+    if [ -f /etc/redhat-release ]; then
+      if [ -f /etc/centos-release ]; then
+        OS=CentOS
+      else
+        OS=RedHat
+      fi
+      OSVER=`rpm -qf /etc/redhat-release --qf="%{VERSION}.%{RELEASE}\n" | awk -F. '{print $1"."$2}'`
+      OSREL=`rpm -qf /etc/redhat-release --qf="%{VERSION}\n"`
+    fi
+  fi
+}
+
+# Check to see if we are on a supported OS.
+discover_os
+if [ "$OS" != RedHat -a "$OS" != CentOS -a "$OS" != Ubuntu ]; then
+  echo "ERROR: Unsupported OS."
+  exit 3
 fi
+
 PROXY=`egrep -h '^ *http_proxy=http|^ *https_proxy=http' /etc/profile.d/*`
 eval $PROXY
 export http_proxy
 export https_proxy
-wget -q https://archive.cloudera.com/director/redhat/${OSREL}/x86_64/director/cloudera-director.repo -O /etc/yum.repos.d/cloudera-director.repo
-yum -y -e1 -d1 install cloudera-director-server cloudera-director-client
+if [ -z $http_proxy ]; then
+  PROXY=`egrep -l 'http_proxy=|https_proxy=' /etc/profile.d/*`
+  . $PROXY
+fi
+
+if [ "$OS" == RedHat -o "$OS" == CentOS ]; then
+  # Because it may have been put there by some other process.
+  if [ ! -f /etc/yum.repos.d/cloudera-director.repo ]; then
+    wget -q https://archive.cloudera.com/director/redhat/${OSREL}/x86_64/director/cloudera-director.repo -O /etc/yum.repos.d/cloudera-director.repo
+  fi
+  yum -y -e1 -d1 install cloudera-director-server cloudera-director-client
+  chkconfig cloudera-director-server on
+elif [ "$OS" == Ubuntu ]; then
+  # Because it may have been put there by some other process.
+  if [ ! -f /etc/apt/sources.list.d/cloudera-director.list ]; then
+    wget -q https://archive.cloudera.com/director/ubuntu/${OSNAME}/amd64/director/cloudera.list -O /etc/apt/sources.list.d/cloudera-director.list
+    curl -s http://archive.cloudera.com/director/ubuntu/${OSNAME}/amd64/director/archive.key | apt-key add -
+  fi
+  apt-get -y -q update
+  apt-get -y -q install cloudera-director-server cloudera-director-client
+  update-rc.d cloudera-director-server defaults
+fi
 if [ ! -f /etc/cloudera-director-server/application.properties-orig ]; then
   cp -p /etc/cloudera-director-server/application.properties /etc/cloudera-director-server/application.properties-orig
 else
@@ -36,5 +82,4 @@ sed -i -e '/lp.encryption.twoWayCipher:/a\
 lp.encryption.twoWayCipher: desede' -e "/lp.encryption.twoWayCipherConfig:/a\
 lp.encryption.twoWayCipherConfig: `python -c 'import base64, os; print base64.b64encode(os.urandom(24))'`" /etc/cloudera-director-server/application.properties
 service cloudera-director-server start
-chkconfig cloudera-director-server on
